requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
pydantic>=2.0.0
python-dotenv>=1.0.0
mcp>=1.0.0
pypdf>=4.0.0
rank-bm25>=0.2.2